CDK updates:
1 - move cdk.json to src/cdk directory

This is for a couple of reasons:
- One fewer file in project root, which I think is A Good Thing
- Makes it easier to have repos with multiple, separate, CDK apps

2 - update lambdaFunction definition for new relative location of app source code

This is necessary because of the above change - CDK is going to be running in a different directory now

3 - move "output" and "requireApproval" CDK settings from package.json to cdk.json

I hadn't read the docs enough to know they could be in cdk.json. Oops. This way is cleaner

4 - Use CDK 'build' step to run TSC instead of 'preCompilation' in NodejsFunction

Getting NodejsFunction to run tsc just for type-checking is problematic, but I had it working in previous version.
I hadn't realized before that the "build" CDK property applied to *all* synth activities, not just "cdk watch", and so putting project-wide type checking in the "build" property, combined with still using NodejsFunction for running esbuild, is a cleaner solution.

This allows cleaning up the workaround in tsconfig.json